Refactor Python emulator into modular structure
Split the monolithic emulator.py file into separate modules for better code organization and maintainability. Each peripheral type (Pin, UART, I2C) now has its own module, making the codebase easier to understand and extend.

Changes:
- Create common.py for shared types (Status enum, UnhandledMessageError)
- Create pin.py for Pin class and digital pin emulation
- Create uart.py for Uart class and UART peripheral emulation
- Create i2c.py for I2C class and I2C controller emulation
- Update emulator.py to only contain DeviceEmulator class
- Update __init__.py to export public API with relative imports
- Update test imports to use new package structure

All tests continue to pass with the new modular structure.
