+++ py/host-emulator/src/__init__.py
@@ -1,0 +1,16 @@
+"""Host emulator for embedded C++ applications."""
+
+from .common import Status, UnhandledMessageError
+from .emulator import DeviceEmulator
+from .i2c import I2C
+from .pin import Pin
+from .uart import Uart
+
+__all__ = [
+    "DeviceEmulator",
+    "Pin",
+    "Uart",
+    "I2C",
+    "Status",
+    "UnhandledMessageError",
+]
